mnt/user-data/outputs/yolo-microservices/gateway/requirements.txt
flask>=3.0
grpcio>=1.60
grpcio-tools>=1.60
protobuf>=4.25
pillow>=10.0
numpy>=1.24
opencv-python-headless>=4.8
